checking if appsaway_app_path_not_console is defined before removing anything

+++ scripts/appsAway_setupCluster.sh
@@ -241,11 +241,16 @@ copy_yaml_files()
 {
   nodes_addr_list=(${APPSAWAY_NODES_ADDR_LIST})
   nodes_username_list=(${APPSAWAY_NODES_USERNAME_LIST})
-  for index in "${!nodes_addr_list[@]}"
-  do
-    log "Removing ${_OS_HOME_DIR}/${nodes_username_list[$index]}/${APPSAWAY_APP_PATH_NOT_CONSOLE} on node ${nodes_addr_list[$index]}..."
-    run_via_ssh_no_folder ${nodes_username_list[$index]} ${nodes_addr_list[$index]} "if [ -d '${_OS_HOME_DIR}/${nodes_username_list[$index]}/${APPSAWAY_APP_PATH_NOT_CONSOLE}' ]; then rm -rf ${_OS_HOME_DIR}/${nodes_username_list[$index]}/${APPSAWAY_APP_PATH_NOT_CONSOLE}; fi"
-  done
+  if [ -v APPSAWAY_APP_PATH_NOT_CONSOLE ]
+  then
+    for index in "${!nodes_addr_list[@]}"
+    do
+      log "Removing ${_OS_HOME_DIR}/${nodes_username_list[$index]}/${APPSAWAY_APP_PATH_NOT_CONSOLE} on node ${nodes_addr_list[$index]}..."
+      run_via_ssh_no_folder ${nodes_username_list[$index]} ${nodes_addr_list[$index]} "if [ -d '${_OS_HOME_DIR}/${nodes_username_list[$index]}/${APPSAWAY_APP_PATH_NOT_CONSOLE}' ]; then rm -rf ${_OS_HOME_DIR}/${nodes_username_list[$index]}/${APPSAWAY_APP_PATH_NOT_CONSOLE}; fi"
+    done
+  else
+    log "APPSAWAY_APP_PATH_NOT_CONSOLE not defined, skipping removal..."
+  fi
   log "creating path ${APPSAWAY_APP_PATH} on master node (this)"
   mkdir -p ${APPSAWAY_APP_PATH}
   for file in ${APPSAWAY_DEPLOY_YAML_FILE_LIST}
